Move DispatcherProvider to core
