+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ErrorCall.kt
@@ -2,8 +2,8 @@ package io.getstream.chat.android.client.api
 
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.errors.ChatError
-import io.getstream.chat.android.client.internal.DispatcherProvider
 import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
