+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/call/ZipCall.kt
@@ -1,8 +1,8 @@
 package io.getstream.chat.android.client.call
 
 import io.getstream.chat.android.client.errors.ChatError
-import io.getstream.chat.android.client.internal.DispatcherProvider
 import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.async
