+++ stream-chat-android-offline/build.gradle
@@ -76,6 +76,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 }
 
 dependencies {
+    implementation project(':stream-chat-android-core')
     api project(":stream-chat-android-client")
 
     // Kotlin
