+++ stream-chat-android-ui-components/build.gradle
@@ -73,6 +73,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 }
 
 dependencies {
+    implementation project(':stream-chat-android-core')
     api project(":stream-chat-android")
 
     implementation Dependencies.kotlinStdLib
