+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/viewmodel/CreateChannelViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.getstream.chat.android.client.ChatClient
-import io.getstream.chat.android.client.internal.DispatcherProvider
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Member
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.models.name
 import io.getstream.chat.android.livedata.ChatDomain
+import io.getstream.chat.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.launch
 
 private val CHANNEL_NAME_REGEX = Regex("^!?[\\w-]*\$")
