+++ stream-chat-android/src/test/java/com/getstream/sdk/chat/utils/TestCoroutineExtension.kt
@@ -2,7 +2,7 @@
 
 package com.getstream.sdk.chat.utils
 
-import io.getstream.chat.android.client.internal.DispatcherProvider
+import io.getstream.chat.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.TestCoroutineScope
 import org.junit.jupiter.api.extension.AfterAllCallback
